requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==24.2
django-simple-history==3.7.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-camel-case==1.4.2
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
filelock==3.14.0
gunicorn==22.0.0
identify==2.5.36
idna==3.7
Markdown==3.6
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.1
platformdirs==4.2.2
pre-commit==3.7.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
PyYAML==6.0.1
requests==2.32.2
requests-oauthlib==2.0.0
ruff==0.4.8
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
urllib3==2.2.1
virtualenv==20.26.2
wheel==0.43.0
